reorg fix. move back logo.svg under /src/.../
