+++ chathaven/public/index.html
@@ -15,7 +15,7 @@
     <meta charSet="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>ChatHaven</title>
-    <link rel="icon" type="image/x-icon" href="%PUBLIC_URL%/logo.svg" />
+    <link rel="icon" type="image/x-icon" href="%PUBLIC_URL%/chathaven/src/assets/logo.svg" />
     <link rel="manifest" href="/manifest.json" />
   </head>
   <body>
