+++ chathaven/src/components/SideBar.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import '../styles/tailwind.CSS/tailwind.css'
-import logo from "../../public/logo.svg";
+import logo from "../assets/logo.svg";
 import more from "../../public/icons/more.svg";
